changes to various exercises
Updated output validation in Tender Change exercise - dynamic value
variable in function only returns denominations with a value > 0 in
change variable.
Commented out validation inputs and outputs in pluralize exercise
because there is a syntax error which causes the page not to render in
the browser. Will need to follow up in class to understand why.
Added 'return' to movieParsing function.  Need to create a validation
still.